#53 Plug kover on multiple modules project
