--- app/app/src/main/kotlin/com/fabien/app/organization/handler/AddOrganizationCommandHandler.kt
@@ -1,42 +1,0 @@
-package com.fabien.app.organization.handler
-
-import arrow.core.NonEmptyList
-import arrow.core.raise.either
-import arrow.core.raise.ensure
-import arrow.core.raise.zipOrAccumulate
-import com.fabien.domain.*
-import com.fabien.domain.handlers.AddOrganizationCommandHandler
-import com.fabien.domain.model.NewOrganization
-import com.fabien.domain.repositories.OrganizationRepository
-import java.lang.Boolean.FALSE
-
-fun addOrganizationCommandHandler(organizationRepository: OrganizationRepository) = AddOrganizationCommandHandler { addCommand ->
-    either {
-        val (name, nationalId, zipCode, country, city, address, active) = addCommand
-        // TODO TRIM field name, zipcode, country, city, address
-        // TODO remove space from nationalID
-        either<NonEmptyList<OrganizationCreationError>, NewOrganization> {
-            zipOrAccumulate(
-                { ensure(!name.isNullOrEmpty()) { MissingName } },
-                { ensure(!nationalId.isNullOrEmpty()) { MissingNationalId } },
-                { ensure(!country.isNullOrEmpty()) { MissingCountry } },
-                // TODO add national Id check
-            ) { _, _, _ ->
-                // TODO Upper case country, city
-                NewOrganization(
-                    name = name!!,
-                    nationalId = nationalId!!,
-                    zipCode = zipCode,
-                    country = country!!,
-                    city = city,
-                    address = address,
-                    active = !FALSE.equals(active),
-                )
-            }
-        }.mapLeft { errors ->
-            OrganizationCreationErrorList(errors)
-        }.bind().let { organization ->
-            organizationRepository.save(organization).bind()
-        }
-    }
-}
